+++ flutter_osm_interface/lib/src/osm_interface.dart
@@ -101,8 +101,9 @@ abstract class MobileOSMPlatform extends OSMPlatform {
   );
 
   Future<void> enableTracking(
-    int idOSM,
-  );
+    int idOSM, {
+    bool stopFollowInDrag = false,
+  });
 
   Future<void> addPosition(
     int idOSM,