+++ ios/Classes/MyMapView.swift
@@ -54,6 +54,8 @@ public class MyMapView: NSObject, FlutterPlatformView, CLLocationManagerDelegate
     var customTiles: [String: Any]? = nil
     var oldCustomTile: CustomTiles? = nil
     var bounds: [Double]? = nil
+    var enableStopFollowInDrag: Bool = false
+    var canSkipFollow: Bool = false
     let urlStyle = "https://github.com/liodali/osm_flutter/raw/dc7424dacd77f4eced626abf64486d70fd03240d/assets/dynamic-styles.zip"
 
     init(_ frame: CGRect, viewId: Int64, channel: FlutterMethodChannel, args: Any?) {
@@ -182,6 +184,7 @@ public class MyMapView: NSObject, FlutterPlatformView, CLLocationManagerDelegate
             result(mapView.position.toGeoPoint())
             break;
         case "trackMe":
+            enableStopFollowInDrag = call.arguments as? Bool ?? false
             trackUserLocation()
             result(200)
             break;
@@ -469,7 +472,7 @@ public class MyMapView: NSObject, FlutterPlatformView, CLLocationManagerDelegate
 
     private func addMarkerManually(call: FlutterMethodCall) {
         let args = call.arguments as! [String: Any]
-        var icon = markerIcon!
+        var icon = markerIcon ?? defaultIcon
         if (args.keys.contains("icon")) {
             let iconArg = args["icon"] as! [String: Any]
             icon = MarkerIconData(image: convertImage(codeImage: iconArg["icon"] as! String), size: iconArg["size"] as! [Int])
@@ -744,6 +747,7 @@ public class MyMapView: NSObject, FlutterPlatformView, CLLocationManagerDelegate
         let args = call.arguments as! [String: Any]
         var points = args["wayPoints"] as! [GeoPoint]
         var roadType = RoadType.car
+        let keepInitial = args["keepInitialGeoPoint"] as! Bool
         switch args["roadType"] as! String {
         case "car":
             roadType = RoadType.car
@@ -758,12 +762,14 @@ public class MyMapView: NSObject, FlutterPlatformView, CLLocationManagerDelegate
             roadType = RoadType.car
             break
         }
-        points.forEach { p in
-            let markers = mapView.markers.filter { m in
-                m.point == p.toLocationCoordinate()
-            }
-            markers.forEach { m in
-                mapView.markerRemove(m)
+        if keepInitial {
+            points.forEach { p in
+                let markers = mapView.markers.filter { m in
+                    m.point == p.toLocationCoordinate()
+                }
+                markers.forEach { m in
+                    mapView.markerRemove(m)
+                }
             }
         }
         if (!roadManager.roads.isEmpty) {
@@ -811,8 +817,9 @@ public class MyMapView: NSObject, FlutterPlatformView, CLLocationManagerDelegate
                 box = route.coordinates?.toBounds()
             }
             var interestGeoPoints: [GeoPointMap]? = nil
+
             if let showMarkerInPOI = args["showMarker"] as? Bool {
-                if (showMarkerInPOI) {
+                if (showMarkerInPOI && !keepInitial) {
                     interestGeoPoints = [GeoPointMap]()
                     let start = self.markersIconsRoadPoint["start"] ?? self.defaultIcon
                     let geoStartM = GeoPointMap(icon: start!, coordinate: points.first!.toLocationCoordinate())
@@ -960,7 +967,15 @@ public class MyMapView: NSObject, FlutterPlatformView, CLLocationManagerDelegate
                 if (canGetLastUserLocation) {
                     canGetLastUserLocation = false
                 }
-                mapView.flyToUserLocation(for: location)
+                if !canSkipFollow && !enableStopFollowInDrag {
+                    mapView.flyToUserLocation(for: location) { [self] end in
+                        if enableStopFollowInDrag {
+                            canSkipFollow = true
+                        }
+
+                    }
+                }
+
 
             }
         } else if (retrieveLastUserLocation) {
@@ -1032,6 +1047,7 @@ public class MyMapView: NSObject, FlutterPlatformView, CLLocationManagerDelegate
             let data: [String: Any] = ["center": point, "bounding": bounding]
             channel.invokeMethod("receiveRegionIsChanging", arguments: data)
         }
+
 
     }