+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 0.42.0 : add api to enableTracking
+* add parameter `enableStopFollow` in `enableTracking`
+* disable follow track user location when user start drag
+* fix drawRoad : add keepInitialGeo in ios side
 ### 0.41.2 : fix dependencies for android sdk 33
 ### 0.41.1 : add `withPosition` and `withUserPosition` as new named constructor for `MapController`
 ### 0.41.0 : change tile in runtime